requirements.txt
yfinance>=0.2.66
pandas>=2.2.3
numpy>=2.1.2
matplotlib>=3.9.2
scikit-learn>=1.7.2
joblib>=1.4.2
platformdirs>=4.3.6
pillow>=11.0.0
